Refactored observables in dedicated files to reduce file complexity and improve readability.
