+++ bgw-core/src/main/kotlin/tools/aqua/bgw/observable/Property.kt
@@ -65,108 +65,3 @@ open class Property<T>(initialValue: T) : ValueObservable<T>() {
 	fun notifyUnchanged(): Unit = notifyChange(boxedValue, boxedValue)
 }
 
-/**
- * A BooleanProperty.
- *
- * @constructor Creates a [BooleanProperty] with given initial value.
- *
- * @param initialValue Initial Value. Default: `false`.
- */
-open class BooleanProperty(initialValue: Boolean = false) : Property<Boolean>(initialValue)
-
-/**
- * An IntegerProperty.
- *
- * @constructor Creates a [IntegerProperty] with given initial value.
- *
- * @param initialValue Initial Value. Default: 0.
- */
-open class IntegerProperty(initialValue: Int = 0) : Property<Int>(initialValue)
-
-/**
- * A DoubleProperty.
- *
- * @constructor Creates a [DoubleProperty] with given initial value.
- *
- * @param initialValue Initial Value. Default: 0.0.
- */
-open class DoubleProperty(initialValue: Number = 0.0) : Property<Double>(initialValue.toDouble())
-
-/**
- * A limited DoubleProperty to a value range. Value will be checked to be in range
- * [lowerBoundInclusive] to [upperBoundInclusive].
- *
- * Therefore [upperBoundInclusive] must be greater or equal to [lowerBoundInclusive].
- * The Range is constant and cannot be altered after object creation.
- *
- * @constructor Creates a [LimitedDoubleProperty] with given bounds and initial value.
- *
- * @param lowerBoundInclusive Lower bound inclusive. Default: -inf.
- * @param upperBoundInclusive Upper bound inclusive. Default: +inf.
- * @param initialValue Initial Value. Default: [lowerBoundInclusive].
- *
- * @throws IllegalArgumentException If a value out of range is set as initialValue.
- */
-open class LimitedDoubleProperty(
-	lowerBoundInclusive: Number = Double.NEGATIVE_INFINITY,
-	upperBoundInclusive: Number = Double.POSITIVE_INFINITY,
-	initialValue: Number = lowerBoundInclusive
-) : Property<Double>(initialValue.toDouble()) {
-	
-	private val lowerBoundInclusive: Double
-	private val upperBoundInclusive: Double
-	
-	init {
-		require(lowerBoundInclusive.toDouble() <= upperBoundInclusive.toDouble()) {
-			"Argument is lower than lower bound for this property."
-		}
-		this.lowerBoundInclusive = lowerBoundInclusive.toDouble()
-		this.upperBoundInclusive = upperBoundInclusive.toDouble()
-		
-		checkBounds(initialValue.toDouble())
-	}
-	
-	/**
-	 * Value of this property.
-	 */
-	override var value: Double
-		get() = super.value
-		set(value) {
-			checkBounds(value)
-			super.value = value
-		}
-	
-	/**
-	 * Overrides value of this property without notifying public listeners.
-	 *
-	 * Only notifies [guiListener].
-	 */
-	override fun setSilent(value: Double) {
-		checkBounds(value)
-		
-		super.setSilent(value)
-	}
-	
-	/**
-	 * Checks whether the given value is in the valid range.
-	 *
-	 * @return `true` if the given value is in the valid range, `false` otherwise.
-	 */
-	private fun checkBounds(value: Double) {
-		require(value >= lowerBoundInclusive) {
-			"Argument is lower than lower bound for this property."
-		}
-		require(value <= upperBoundInclusive) {
-			"Argument is higher than upper bound for this property."
-		}
-	}
-}
-
-/**
- * A StringProperty.
- *
- * @constructor Creates a [StringProperty] with given initial value.
- *
- * @param initialValue Initial Value. Default: Empty string.
- */
-open class StringProperty(initialValue: String = "") : Property<String>(initialValue)
