+++ bgw-core/src/main/kotlin/tools/aqua/bgw/observable/ValueObservable.kt
@@ -20,152 +20,6 @@
 package tools.aqua.bgw.observable
 
 /**
- * Observable interface for observable Properties.
- */
-@FunctionalInterface
-fun interface IObservable {
-	/**
-	 * Indicates property update.
-	 */
-	fun update()
-}
-
-/**
- * Observable interface for observable properties with values.
- */
-@FunctionalInterface
-fun interface IValueObservable<T> {
-	/**
-	 * Indicates property update.
-	 *
-	 * @param oldValue Old value of property.
-	 * @param newValue New value of property.
-	 */
-	fun update(oldValue: T, newValue: T)
-}
-
-/**
- * Basic observable.
- *
- * @constructor Empty constructor.
- */
-@Suppress("UnnecessaryAbstractClass")
-abstract class Observable {
-	/**
-	 * Holds all listeners on this property.
-	 */
-	private val listeners: MutableList<IObservable> = mutableListOf()
-	
-	/**
-	 * Used by renderer to listen on important properties for visualization.
-	 */
-	private var guiListenerHandler: IObservable? = null
-	
-	/**
-	 * Used by renderer to listen on important properties for visualization.
-	 */
-	internal var guiListener: (() -> Unit)? = null
-		set(value) {
-			guiListenerHandler = if (value == null) null else IObservable(value)
-		}
-	
-	/**
-	 * Used by BGW framework containers to manage children.
-	 *
-	 * Should only be set by direct parent.
-	 */
-	private var internalListenerHandler: IObservable? = null
-	
-	/**
-	 * Used by BGW framework containers to manage children.
-	 *
-	 * Should only be set by direct parent.
-	 */
-	internal var internalListener: (() -> Unit)? = null
-		set(value) {
-			internalListenerHandler = if (value == null) null else IObservable(value)
-		}
-	
-	/**
-	 * Sets [guiListener] and calls [IObservable.update].
-	 *
-	 * @param listener Listener to add and notify.
-	 */
-	internal fun setGUIListenerAndInvoke(listener: () -> Unit) {
-		guiListener = listener
-		guiListenerHandler?.update()
-	}
-	
-	/**
-	 * Sets [internalListener] and calls [IObservable.update].
-	 *
-	 * @param listener Listener to add and notify.
-	 */
-	internal fun setInternalListenerAndInvoke(listener: () -> Unit) {
-		internalListener = listener
-		internalListenerHandler?.update()
-	}
-	
-	/**
-	 * Adds a [listener] and calls [IObservable.update] on this new listener.
-	 *
-	 * @param listener Listener to add and notify.
-	 */
-	fun addListenerAndInvoke(listener: IObservable) {
-		listeners.add(listener)
-		listener.update()
-	}
-	
-	/**
-	 * Adds a [listener] silently.
-	 *
-	 * @param listener Listener to add
-	 */
-	fun addListener(listener: IObservable) {
-		listeners.add(listener)
-	}
-	
-	/**
-	 * Removes a [listener].
-	 *
-	 * @param listener Listener to remove.
-	 *
-	 * @return `true` if the listener has been successfully removed, `false` if it was not found.
-	 */
-	fun removeListener(listener: IObservable): Boolean = listeners.remove(listener)
-	
-	/**
-	 * Removes all listeners.
-	 */
-	fun clearListeners() {
-		listeners.clear()
-	}
-	
-	/**
-	 * Notifies [guiListener] by calling [IObservable.update].
-	 */
-	internal fun notifyGUIListener() {
-		guiListenerHandler?.update()
-	}
-	
-	/**
-	 * Notifies [internalListener] by calling [IObservable.update].
-	 */
-	internal fun notifyInternalListener() {
-		internalListenerHandler?.update()
-	}
-	
-	/**
-	 * Notifies all [listeners] by calling [IObservable.update].
-	 */
-	fun notifyChange() {
-		listeners.forEach { it.update() }
-		internalListenerHandler?.update()
-		guiListenerHandler?.update()
-	}
-}
-
-/**
  * Basic observable with value.
  *
  * @constructor Empty constructor.
